+++ inc/class.environments.php
@@ -8,7 +8,7 @@
  *
 **/
 
-namespace R\AdminUtils;
+namespace RH\AdminUtils;
 
 if ( ! defined( 'ABSPATH' ) ) exit; // Exit if accessed directly