+++ rh-admin-utils.php
@@ -1,13 +1,13 @@
 <?php
 /**
  * Plugin Name: RH Admin Utilities
- * Version: 1.2.2
+ * Version: 1.2.3
  * Author: Rasso Hilber
  * Description: Admin Utilities for WordPress. Removes plugin ads, adds custom buttons to the admin bar (publish, clear cache), allows editors to add users (except administrators).
  * Author URI: https://rassohilber.com
 **/
 
-namespace R\AdminUtils;
+namespace RH\AdminUtils;
 
 if ( ! defined( 'ABSPATH' ) ) exit; // Exit if accessed directly
 
@@ -211,7 +211,7 @@ require_once(__DIR__ . '/inc/class.remove-ads.php');
 require_once(__DIR__ . '/inc/class.admin-bar-publish-button.php');
 require_once(__DIR__ . '/inc/class.misc.php');
 require_once(__DIR__ . '/inc/class.environments.php');
-require_once(__DIR__ . '/inc/class.editor-can-update.php');
+require_once(__DIR__ . '/inc/class.editor-in-chief.php');
 
 /**
  * Initialize util classes
@@ -222,4 +222,4 @@ RemoveAds::getInstance();
 AdminBarPublishButton::getInstance();
 Misc::getInstance();
 Environments::getInstance();
-EditorCanUpdate::getInstance();
+EditorInChief::getInstance();